purchase - sales html th:if added, column names changed

+++ src/main/java/com/cydeo/accountingsimplified/dto/InvoiceDto.java
@@ -26,7 +26,7 @@ public class InvoiceDto {
     private ClientVendorDto clientVendor;
 
     private BigDecimal price;
-    private Integer tax;
+    private BigDecimal tax;
     private BigDecimal total;
 
 }
